requirements.txt
# --- Core RAG Framework ---
langchain
langchain-community
langchain-core
langchain-openai
langchain-huggingface
langchain-chroma
langchain-ollama

# --- Vector Database ---
chromadb

# --- Local AI Engines ---
sentence-transformers  # Required for HuggingFaceEmbeddings

# --- Document Loaders ---
# pypdf is not good with tables and complex layouts, use pdfplumber instead
pdfplumber
unstructured

# --- Utilities ---
python-dotenv
tiktoken
numpy
pandas
